Renamed plugin to 'Autocast Utilities'. Refactored files accordingly.
