+++ README.md
@@ -1,8 +1,7 @@
-# Skill Level Notifications
-This plugin aims to notify the client when skill levels fall below a define threshold.
+# Autocast Utilities
+This plugin aims to notify the client when the player's magic level falls below the level 
+required for the autocast spell.
 
-The first feature to be implemented is to notify the client when the player's magic level
-falls below what is required for the currently selected autocasted spell. This provides the
-player a chance to restore their magic level before attempting to autocast, which would
-normally mean that the autocasted spell would become deselected and must be
-reselected in the autocast menu. 
+This provides the player a chance to restore their magic level before attempting to 
+autocast again, which would normally mean that the autocasted spell would become 
+deselected and must be reselected in the autocast menu.
