+++ runelite-plugin.properties
@@ -1,6 +1,6 @@
-displayName=SkillLevelNotifications
+displayName=AutocastUtilities
 author=JC Arbelbide (jcarbelbide on github)
 support=
-description=Notifies client when skills fall below a threshold.
-tags=skill,notifier,notifications,mage,magic,reduced,reduction,level,drain,autocast,cast
-plugins=com.skilllevelnotifications.SkillLevelNotificationsPlugin
+description=Notifies client when the player's magic level falls below the level required for the autocast spell.
+tags=notifier,notifications,mage,magic,reduced,reduction,level,drain,autocast,cast,utilities
+plugins=com.autocastutilities.AutocastUtilitiesPlugin
